testing/web-platform/tests/uievents/click/click_events_on_input.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Clicking with primary vs non-primary buttons</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-actions.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
  </head>
  <body>
    <h1>Clicking on input type=text element when placeholder changes</h1>
    <input id="target" onfocus="this.placeholder = ++focusCount;" placeholder="initial">
    <input id="other">
    <script>
    var focusCount = 0;
    var target = document.querySelector('#target');
    document.addEventListener('contextmenu', event => { event.preventDefault(); });

    var test_click = async_test("Test click and auxclick on input element");

    async function testClick(type, mouseButton) {
      var actions_promise;
      return new Promise((resolve) => {
        target.addEventListener(type, event => {
          event.preventDefault();
          test_click.step(() => {
            assert_equals(event.type, type, 'Should have got an event.');
          });
          // Make sure the test finishes after all the input actions are completed.
          actions_promise.then(resolve);
        }, {once: true});

        // Inject mouse click events.
        var actions = new test_driver.Actions();
        document.getElementById("other").focus();
        var bounds = target.getBoundingClientRect();
        actions_promise =
          actions.pointerMove(Math.floor(bounds.width / 2),
                              Math.floor(bounds.height / 2),
                              {origin: target})
                 .pointerDown({button: mouseButton})
                 .pointerUp({button: mouseButton})
                 .send();
      });
    }

    async function testClicks() {
      var buttonType = test_driver.Actions.prototype.ButtonType;
      await testClick("click", buttonType.LEFT);
      await testClick("auxclick", buttonType.MIDDLE);
      await testClick("auxclick", buttonType.RIGHT);
      test_click.done();
    }

    testClicks();
    </script>
  </body>
</html>
