LayoutTests/imported/w3c/web-platform-tests/uievents/click/click_events_on_input-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Clicking on input type=text element when placeholder changes



Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Test click and auxclick on input element
